Moved file and link content classes and interfaces into their own subfolders and namespaces
