+++ AuthoringTool/Mapping/EntityPersistEntityMappingProfile.cs
@@ -2,6 +2,8 @@ using AutoMapper;
 using AutoMapper.EquivalencyExpression;
 using BusinessLogic.Entities;
 using BusinessLogic.Entities.LearningContent;
+using BusinessLogic.Entities.LearningContent.FileContent;
+using BusinessLogic.Entities.LearningContent.LinkContent;
 using PersistEntities;
 using PersistEntities.LearningContent;
