+++ BusinessLogic/Entities/LearningContent/FileContent/IFileContent.cs
@@ -1,0 +1,7 @@
+namespace BusinessLogic.Entities.LearningContent.FileContent;
+
+public interface IFileContent : ILearningContent
+{
+    string Type { get; set; }
+    string Filepath { get; set; }
+}
