--- BusinessLogic/Entities/LearningContent/IFileContent.cs
@@ -1,7 +1,0 @@
-namespace BusinessLogic.Entities.LearningContent;
-
-public interface IFileContent : ILearningContent 
-{
-    string Type { get; set; }
-    string Filepath { get; set; }
-}
