+++ BusinessLogicTest/Commands/Element/CreateLearningElementInSlotUt.cs
@@ -1,6 +1,7 @@
 using BusinessLogic.Commands.Element;
 using BusinessLogic.Entities;
 using BusinessLogic.Entities.LearningContent;
+using BusinessLogic.Entities.LearningContent.FileContent;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
 using Shared;
