+++ Presentation/Components/Forms/Content/AddLinkForm.razor
@@ -2,7 +2,7 @@
 @using Presentation.PresentationLogic.API
 @using Presentation.PresentationLogic.AuthoringToolWorkspace
 @using Presentation.PresentationLogic.LearningContent
-@using BusinessLogic.Entities.LearningContent
+@using BusinessLogic.Entities.LearningContent.LinkContent
 @using AutoMapper
 @using Microsoft.Extensions.Localization
 @using System.Diagnostics.CodeAnalysis
